+++ components/brave_new_tab_ui/components/default/notification/index.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import styled from '../../../../theme'
+import styled from 'brave-ui/theme'
 
 export const SiteRemovalNotification = styled<{}, 'header'>('header')`
   font-family: ${p => p.theme.fontFamily.heading};
