+++ components/brave_new_tab_ui/components/default/page/index.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import styled, { keyframes } from '../../../../theme'
+import styled, { keyframes } from 'brave-ui/theme'
 
 const fadeIn = keyframes`
   from {
