+++ components/brave_new_tab_ui/containers/newTab/settings.tsx
@@ -4,9 +4,9 @@
 
 import * as React from 'react'
 
-import { SettingsMenu, SettingsRow, SettingsText, SettingsTitle, SettingsWrapper } from 'brave-ui/features/newTab/default'
+import { SettingsMenu, SettingsRow, SettingsText, SettingsTitle, SettingsWrapper } from '../../components/default'
 
-import { Toggle } from 'brave-ui/features/newTab/toggle'
+import { Toggle } from '../../components/toggle'
 
 import { getLocale } from '../../../common/locale'
