+++ components/brave_new_tab_ui/containers/privateTab/index.tsx
@@ -5,7 +5,7 @@
 import * as React from 'react'
 
 // Feature-specific components
-import { Page, PageWrapper } from 'brave-ui/features/newTab/private'
+import { Page, PageWrapper } from '../../components/private'
 
 // Components group
 import PrivateTab from './privateTab'
