+++ components/brave_new_tab_ui/containers/privateTab/qwantTab.tsx
@@ -15,7 +15,7 @@ import {
   TorImage,
   Separator,
   FakeButton
-} from 'brave-ui/features/newTab/private'
+} from '../../components/private'
 import TorContent from './torContent'
 // Helpers
 import { getLocale } from '../../../common/locale'
