+++ components/brave_new_tab_ui/containers/privateTab/torTab.tsx
@@ -21,7 +21,7 @@ import {
   Separator,
   FakeButton,
   Link
-} from 'brave-ui/features/newTab/private'
+} from '../../components/private'
 
 // Helpers
 import { getLocale } from '../../../common/locale'
